requirements.txt
python-dateutil>=2.8.2
pytest>=7.0
pillow>=9.0
piexif>=1.1.3
